GenericValue: avoid memset/memcpy(this,...)
To avoid writing outside of the current GenericValue object and
to follow the C++ standard more closely, this patch drops the
memset/memcpy(this,...) occurences in GenericValue in favour of
explicit initialisations/assignments of the data_ and flag_ members.

This should fix http://code.google.com/p/rapidjson/issues/detail?id=11.